add wx MonkeyPatches to package
