+++ source/monkeyPatches/__init__.py
@@ -3,6 +3,11 @@
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
 
+from . import wxMonkeyPatches
+
+
+applyWxMonkeyPatches = wxMonkeyPatches.apply
+
 
 def applyMonkeyPatches():
 	# Apply several monkey patches to comtypes
